requirements.txt
docker
argparse
gitpython
pyyaml ~=5.1